külön oldal egy terméknek

+++ database/migrations/2025_02_13_103924_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("name");
             $table->text("description");
-            $table->decimal("price");
+            $table->integer("price");
             $table->string("image");
             $table->integer("stock_quantity");
             $table->foreignId("category_id")->constrained();
